Fix setState flow
